+++ Project1/ConsoleApp1/Operacion.cs
@@ -26,7 +26,7 @@ namespace ConsoleApp1
         public virtual int sumar() { return 0; }
         public virtual int restar() { return 0; }
         public virtual int multiplicar() { return 0; }
-        public virtual double dividir() { return 0; }
+        public virtual decimal dividir() { return 0; }
 
     }
 }
